assignment 3 commit
In this assignment you will work with a design created from the requirements in assignment 2. You will be presented with the design documentation, the partial implementation, and automatic test cases.
Your first task in this assignment is to implement the missing parts of the code-base according to the automatic unit tests. When the tests are passing you are done.
The second task in this assignment is to reflect on your design as done in assignment 2 and write a short report with your thoughts.
As usual you are encouraged to commit and push often. We want to see an evolution of the solution not a big bang commit at the end.

o build confidence in that our system works as expected after successfull compilation we can run code that test the classes on our own code. Such test code is called unit tests, often we use some framework that can automatically execute our tests and report back any issues. In such cases we talk about automatic unit tests. We will not go deeper in to this for the assignment other than that you will get a number of unit tests and these should be passing when handing in. Basically you take a look at the first test case, change/add to the implementation so that the test case works (you do not change the test case).
The test are located in app/src/test and usually follows the same package structure as the main source code. Note that this means that test code is located in the same package as the main code (even if they have some parent folder that is not the same). This makes it possible for test code to access anything that is declared as protected and not only public public.
The unit tests are of course executed as part of the automatic pipeline and when build using gradlew.